moved the created component to the app/login directory
