+++ src/app/login/login.module.css
@@ -1,4 +1,4 @@
-/* LoginPage*/
+/* Login*/
 
 .container {
     display: flex;
